requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
pyyaml>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
loguru>=0.7.0

# Crawling
crawl4ai>=0.2.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# Data processing and analysis
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0

# LLM Integration
openai>=1.0.0

# Visualization
plotly>=5.15.0
streamlit-aggrid>=0.3.4

# Development tools
ruff>=0.1.0
mypy>=1.5.0
pytest>=7.4.0
pre-commit>=3.4.0

# Optional utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
